+++ src/main/java/project/reviewing/evaluation/command/domain/Evaluation.java
@@ -1,4 +1,4 @@
-package project.reviewing.evaluation.domain;
+package project.reviewing.evaluation.command.domain;
 
 import lombok.AccessLevel;
 import lombok.Getter;
