+++ src/main/java/project/reviewing/evaluation/presentation/EvaluationController.java
@@ -5,10 +5,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.web.bind.annotation.*;
 import project.reviewing.auth.presentation.AuthenticatedMember;
-import project.reviewing.evaluation.application.EvaluationService;
-import project.reviewing.evaluation.application.response.EvaluationsForReviewerResponse;
-import project.reviewing.evaluation.application.response.SingleEvaluationResponse;
+import project.reviewing.evaluation.command.application.EvaluationService;
+import project.reviewing.evaluation.command.application.response.EvaluationsForReviewerResponse;
+import project.reviewing.evaluation.query.application.response.SingleEvaluationResponse;
 import project.reviewing.evaluation.presentation.request.EvaluationCreateRequest;
+import project.reviewing.evaluation.query.application.EvaluationQueryService;
 
 import javax.validation.Valid;
 
@@ -17,6 +18,7 @@ import javax.validation.Valid;
 public class EvaluationController {
 
     private final EvaluationService evaluationService;
+    private final EvaluationQueryService evaluationQueryService;
 
     @PostMapping("/reviewers/{reviewer-id}/evaluations")
     public void createEvaluation(
@@ -29,7 +31,7 @@ public class EvaluationController {
 
     @GetMapping("/evaluations/{review-id}")
     public SingleEvaluationResponse readSingleEvaluation(@PathVariable("review-id") final Long reviewId) {
-        return evaluationService.findSingleEvaluationByReviewId(reviewId);
+        return evaluationQueryService.findSingleEvaluationByReviewId(reviewId);
     }
 
     @GetMapping("reviewers/{reviewer-id}/evaluations")
