+++ src/test/java/project/reviewing/integration/IntegrationTest.java
@@ -10,8 +10,8 @@ import org.springframework.stereotype.Repository;
 import project.reviewing.auth.domain.RefreshTokenRepository;
 import project.reviewing.common.util.ReviewingTime;
 import project.reviewing.common.util.Time;
-import project.reviewing.evaluation.domain.Evaluation;
-import project.reviewing.evaluation.domain.EvaluationRepository;
+import project.reviewing.evaluation.command.domain.Evaluation;
+import project.reviewing.evaluation.command.domain.EvaluationRepository;
 import project.reviewing.member.command.domain.Member;
 import project.reviewing.member.command.domain.MemberRepository;
 import project.reviewing.member.command.domain.Reviewer;
